fix(DataType): Fixed wrong isSameDataType for VARSIZED

+++ nes-data-types/include/DataTypes/DataType.hpp
@@ -50,7 +50,7 @@ struct DataType final
     template <class T>
     [[nodiscard]] bool isSameDataType() const
     {
-        if (this->type == Type::VARSIZED && std::is_same_v<std::remove_cvref_t<T>, std::uint32_t>)
+        if (this->type == Type::VARSIZED && std::is_same_v<std::remove_cvref_t<T>, std::uint64_t>)
         {
             return true;
         }
